Finished

+++ audiodist/views.py
@@ -1,27 +1,70 @@
 from django.shortcuts import render, redirect, get_object_or_404
-from django.http import Http404, JsonResponse, HttpResponse
+from django.http import Http404, JsonResponse, HttpResponse, HttpResponseForbidden, HttpResponseNotAllowed
 from django.template.context_processors import request
+from django.db.models import Q
+from django.contrib.auth import authenticate, login, logout
+from django.contrib import messages
+from django.forms.models import model_to_dict
 
 from .forms import *
 
 # Create your views here.
 def all_views (request):
     try:
-        artist = Artist.objects.get(account=request.user)
+        logged_in_artist = Artist.objects.get(account=request.user)
     except:
-        artist = None
-    return { 'artist': artist }
+        logged_in_artist = None
+
+    return { 'logged_in_artist': logged_in_artist, 'user': request.user }
+
 
 def index_view (request):
-    return render(request, 'audiodist/index.html', {})
+    return render(request, 'audiodist/index.html', { 'artists': Artist.objects.all() })
+
+
+def search_view (request):
+    if request.method == 'GET':
+        query = request.GET.get('search')
+
+        if not query:
+            return redirect('index')
+
+        song_results = Song.objects.filter(Q(title__icontains=query))
+        collection_results = Collection.objects.filter(Q(title__icontains=query) | Q(description__icontains=query))
+        artist_results = list(Artist.objects.filter(Q(name__icontains=query) | Q(description__icontains=query)))
+
+        accounts = Account.objects.filter(is_superuser=False)
+        username_results = [ Artist.objects.get(account=account) for account in accounts.filter(Q(username__icontains=query)) ]
+
+        song_results_count = len(list(filter(lambda s: request.user == s.artist.account or s.is_public, song_results)))
+        collection_results_count = len(list(filter(lambda c: request.user == c.artist.account or c.is_public, collection_results)))
+    elif request.method == 'POST':
+        song_slug = request.POST.get('song_slug')
+        song = get_object_or_404(Song.objects.filter(slug=song_slug))
+        song.plays += 1
+        song.save()
+
+        return JsonResponse({ 'plays': song.plays })
+
+    return render(request, 'audiodist/search.html', { 'query': query, 'results': {
+        'songs': song_results,
+        'collections': collection_results,
+        'artists': list(set(artist_results + username_results)),
+        'tracks': CollectionTrack.objects.all(),
+        'num_results': song_results_count  + collection_results_count + len(list(set(artist_results + username_results)))
+    } })
 
 
 def register_view (request):
+    if request.user.is_authenticated:
+        return redirect('artist', request.user.username)
+
     if request.method == 'POST':
         form = CreateAccountForm(request.POST)
 
         if form.is_valid():
-            form.save()
+            registered_user = form.save()
+            login(request, registered_user)
             return redirect('artist', form.cleaned_data.get('username'))
     else:
         form = CreateAccountForm()
@@ -30,7 +73,34 @@ def register_view (request):
 
 
 def login_view (request):
+    if request.user.is_authenticated:
+        return redirect('artist', request.user.username)
+
+    if request.method == 'POST':
+        email = request.POST.get('email')
+        password = request.POST.get('password')
+
+        account = authenticate(request, username=email, password=password)
+
+        if account is not None:
+            login(request, account)
+            return redirect('artist', account.username)
+        else:
+            messages.info(request, 'Invalid username or password')
+            return redirect('login')
+
     return render(request, 'audiodist/login.html', {})
+
+
+def logout_view (request):
+    if request.method == 'GET':
+        return Http404
+
+    if request.method == 'POST':
+        logout(request)
+        return redirect('index')
+    
+    return JsonResponse({})
 
 
 def artist_view (request, username):
@@ -61,6 +131,13 @@ def artist_view (request, username):
                             follower = Follower.objects.create(of=artist, by=user_artist)
                             follower.save()
                             return JsonResponse({ 'follow': True, "new_num_followers": artist.num_followers })
+                    if request.POST.get('action') == 'play':
+                        song_slug = request.POST.get('song_slug')
+                        song = get_object_or_404(Song.objects.filter(artist=artist), slug=song_slug)
+                        song.plays += 1
+                        song.save()
+
+                        return JsonResponse({ 'plays': song.plays })
                 else:
                     collection_slug = request.POST.get('collection_slug')
                     song_slug = request.POST.get('song_slug')
@@ -80,13 +157,16 @@ def artist_view (request, username):
         raise Http404('User not found')
 
     return render(request, 'audiodist/artist.html', { 'artist': artist, 'songs': artist_songs, 
-    'collections': artist_collections, 'tracks': CollectionTrack.objects.all(), 
+    'collections': artist_collections, 'tracks': CollectionTrack.objects.all(),
     'is_authorized': request.user == artist.account, 'is_user_following': request.user in users_of_followers })
 
 
 def artist_edit_view (request, username):
     if username in [ account.username for account in Account.objects.all() ]:
         artist = Artist.objects.get(account=Account.objects.get(username=username))
+
+        if request.user != artist.account:
+            return HttpResponseForbidden('<h1>You cannot access this page</h1>')
         
         if request.method == 'POST':
             form = ArtistEditForm(request.POST, request.FILES, instance=artist)
@@ -109,6 +189,9 @@ def artist_edit_view (request, username):
 def artist_delete_view (request, username):
     if username in [ account.username for account in Account.objects.all() ]:
         artist = Artist.objects.get(account=Account.objects.get(username=username))
+
+        if request.user != artist.account:
+            return HttpResponseForbidden('<h1>You cannot access this page</h1>')
         
         if request.method == 'POST':
             artist.account.delete()
@@ -121,7 +204,7 @@ def artist_delete_view (request, username):
 
 def song_create_view (request):
     if not request.user.is_authenticated:
-        return redirect('index')
+        return redirect('login')
 
     if request.method == 'POST':
         form = SongForm(request.POST, request.FILES)
@@ -138,7 +221,12 @@ def song_view (request, username, song_slug):
     if username in [ account.username for account in Account.objects.all() ]:
         artist = Artist.objects.get(account=Account.objects.get(username=username))
         song = get_object_or_404(Song.objects.filter(artist=artist), slug=song_slug)
-        collections = Collection.objects.filter(artist=artist)
+        collections = list(filter(lambda c: CollectionTrack.objects.filter(collection=c, song=song).exists(), Collection.objects.filter(artist=artist)))
+
+        comments = Comment.objects.filter(to=song)
+
+        if request.user != artist.account and not song.is_public:
+            return HttpResponseNotAllowed('Access to song not allowed')
 
         if request.method == 'POST':
             if 'action' in request.POST:
@@ -148,7 +236,7 @@ def song_view (request, username, song_slug):
                     return JsonResponse({})
                 if request.POST.get('action') == 'add-to-collection':
                     collection_slug = request.POST.get('collection_slug')
-                    collection = collections.get(slug=collection_slug)
+                    collection = Collection.objects.filter(artist=artist).get(slug=collection_slug)
                     
                     try:
                         track = CollectionTrack.objects.get(collection=collection, song=song)
@@ -156,16 +244,34 @@ def song_view (request, username, song_slug):
                     except:
                         track = CollectionTrack.objects.create(collection=collection, song=song)
                         track.save()
+                if request.POST.get('action') == 'play':
+                    song.plays += 1
+                    song.save()
+
+                    return JsonResponse({ 'plays': song.plays })
+            else:
+                form = CommentForm(request.POST)
+
+                if form.is_valid():
+                    comment = form.save(Artist.objects.get(account=request.user), song)
+                    return JsonResponse({ 'comment': { 'content': comment.content, 'name': comment.by.name, 
+                    'username': comment.by.account.username, 'time': datetime.datetime.now().strftime('%B %d, %Y %-I:%M %p') } })
+        else:
+            form = CommentForm()
     else:
         raise Http404('User not found')
 
-    return render(request, 'audiodist/song.html', { 'artist': artist, 'song': song, 'collections': collections })
+    return render(request, 'audiodist/song.html', { 'artist': artist, 'song': song, 'collections': collections, 
+    'form': form, 'comments': comments })
 
 
 def song_edit_view (request, username, song_slug):
     if username in [ account.username for account in Account.objects.all() ]:
         artist = Artist.objects.get(account=Account.objects.get(username=username))
         song = get_object_or_404(Song.objects.filter(artist=artist), slug=song_slug)
+
+        if request.user != artist.account:
+            return HttpResponseForbidden("<h1>You cannot access this page</h1>")
 
         if request.method == 'POST':
             form = SongEditForm(request.POST, request.FILES, instance=song)
@@ -204,7 +310,7 @@ def song_delete_view (request, username, song_slug):
 
 def collection_create_view (request):
     if not request.user.is_authenticated:
-        return redirect('index')
+        return redirect('login')
 
     if request.method == 'POST':
         form = CollectionForm(request.POST, request.FILES)
@@ -223,6 +329,9 @@ def collection_view (request, username, collection_slug):
         artist = Artist.objects.get(account=Account.objects.get(username=username))
         collection = get_object_or_404(Collection.objects.filter(artist=artist), slug=collection_slug)
         songs = enumerate([ track.song for track in CollectionTrack.objects.filter(collection=collection) ], 1)
+
+        if request.user != artist.account and not collection.is_public:
+            return HttpResponseNotAllowed('Access to collection not allowed')
 
         if request.method == 'POST':
             if 'action' in request.POST:
@@ -245,6 +354,9 @@ def collection_edit_view (request, username, collection_slug):
     if username in [ account.username for account in Account.objects.all() ]:
         artist = Artist.objects.get(account=Account.objects.get(username=username))
         collection = get_object_or_404(Collection.objects.filter(artist=artist), slug=collection_slug)
+
+        if request.user != artist.account:
+            return HttpResponseForbidden("<h1>You cannot access this page</h1>")
 
         if request.method == 'POST':
             form = CollectionForm(request.POST, request.FILES, instance=collection)
